Implement a fairly rough ZZT-OOP interpreter
This is a very minimal interpreter for the ZZT-OOP language. Its only
purpose is to make it possible to run a #give/#take permutation in Ruby,
to make it easier to run statistical tests on the output. As such, it
only supports three instructions:

- #give counter amount
- #take counter amount else_command
- #labelname (jump to :labelname)

Both the parser and interpreter make a basic effort to catch invalid
code, but these checks are not exhaustive. If your ZZT-OOP passes this
parser/interpreter, that's no guarantee that ZZT will also accept it.